+++ main.go
@@ -8,11 +8,13 @@ import (
 	"os"
 )
 
+const version = "0.3"
+
 const helpText = `This is a ccache HTTP(S) storage helper, usually started automatically by ccache
 when needed. More information here: https://ccache.dev/storage-helpers.html
 
 Project: https://github.com/ccache/ccache-storage-http-go
-Version: 0.3
+Version: ` + version + `
 `
 
 func main() {